# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Aryan Landmark Solutions API
  version: 1.0.0
  description: |
    Production REST API for Aryan Landmark Solutions.

    PostgreSQL contains six tables. All property photos, property documents
    and inquiry call recordings use one unified `attachments` resource.

    Cloudflare R2 stores file binaries. PostgreSQL stores metadata and the R2
    object key.
servers:
  - url: /api/v1
...
paths:
...
  /attachments/upload-url:
    post:
      tags: [Attachments]
      summary: Create a short-lived Cloudflare R2 upload authorization
      description: |
        Unified file upload endpoint.

        Use `attachmentType=PHOTO` for property or inquiry photos,
        `attachmentType=DOCUMENT` for 7/12 and other property/inquiry
        documents, `attachmentType=RECORDING` for inquiry call recordings.

        Relationship rules:
        - RECORDING: inquiryId required, propertyId must be absent.
        - PHOTO / DOCUMENT: exactly one of propertyId or inquiryId is
          required (never both). propertyId is the original property-photo
          flow; inquiryId supports the lightweight ADMIN-call-capture flow
          (Phase 13B) where photos/documents are gathered before a
          property record exists.

        The backend validates resource ownership/authorization and returns
        a short-lived upload URL and generated R2 object key.
      security:
        - applicationBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAttachmentUploadRequest'
      responses:
        '200':
          description: R2 upload authorization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '413':
          description: File too large
  /attachments:
    post:
      tags: [Attachments]
      summary: Finalize an uploaded attachment
      description: |
        The client calls this endpoint after uploading the binary to R2.

        The backend verifies the expected object and creates one row in the
        `attachments` table. Same propertyId/inquiryId relationship rules
        as POST /attachments/upload-url.
      security:
        - applicationBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinalizeAttachmentRequest'
      responses:
        '201':
          description: Attachment metadata created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttachmentResponse'
        '400':
          $ref: '#/components/responses/ValidationError'

    get:
      tags: [Attachments]
      summary: List attachments
      description: |
        Filter by propertyId, inquiryId and attachmentType.
      security:
        - applicationBearer: []
      parameters:
        - name: propertyId
          in: query
          schema:
            type: string
            format: uuid
        - name: inquiryId
          in: query
          schema:
            type: string
            format: uuid
        - name: attachmentType
          in: query
          schema:
            $ref: '#/components/schemas/AttachmentType'
      responses:
        '200':
          description: Attachments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttachmentListResponse'
  /attachments/{attachmentId}/download-url:
    get:
      tags: [Attachments]
      summary: Get a short-lived, secure view/download URL for an attachment
      description: |
        Phase 27. The R2 bucket is private (no public r2.dev access, no
        public base URL) — this is the only supported way for an
        authenticated client to obtain a working URL for a photo,
        document, or recording. Verifies the attachment exists, resolves
        its object key from the stored `r2ObjectKey` (never from client
        input — there is no way to request an arbitrary object key
        through this endpoint), and returns a presigned GET URL valid for
        `expiresIn` seconds (300s — short, viewing-session-scoped, not
        meant to be cached or reused beyond one load/open action).
        Same authorization as every other attachment/property read
        (ADMIN or EMPLOYEE, no additional per-resource ownership check).
      security:
        - applicationBearer: []
      parameters:
        - $ref: '#/components/parameters/AttachmentId'
      responses:
        '200':
          description: A short-lived presigned URL for viewing/downloading this attachment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttachmentDownloadUrlResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          description: Cloudflare R2 is not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /attachments/{attachmentId}:
    delete:
      tags: [Attachments]
      summary: Delete attachment
      description: |
        Deletes the attachment metadata and performs the corresponding
        authorized Cloudflare R2 object deletion.
      security:
        - applicationBearer: []
      parameters:
        - $ref: '#/components/parameters/AttachmentId'
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    applicationBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT

    masterBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    AttachmentId:
      name: attachmentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    AttachmentType:
      type: string
      enum:
        - PHOTO
        - DOCUMENT
        - RECORDING
    DocumentType:
      type: string
      enum:
        - SEVEN_TWELVE
        - SALE_DEED
        - PROPERTY_CARD
        - NOC
        - OTHER
...
    CreateAttachmentUploadRequest:
      type: object
      required:
        - attachmentType
        - fileName
        - mimeType
        - fileSizeBytes
      properties:
        attachmentType:
          $ref: '#/components/schemas/AttachmentType'
        propertyId:
          type: string
          format: uuid
          nullable: true
          description: |
            Required for PHOTO/DOCUMENT unless inquiryId is set instead
            (mutually exclusive). Must be absent for RECORDING.
        inquiryId:
          type: string
          format: uuid
          nullable: true
          description: |
            Required for RECORDING. For PHOTO/DOCUMENT, may be used
            instead of propertyId during lightweight inquiry capture
            (Phase 13B) — mutually exclusive with propertyId.
        documentType:
          $ref: '#/components/schemas/DocumentType'
        fileName:
          type: string
        mimeType:
          type: string
        fileSizeBytes:
          type: integer
          minimum: 1
        isPrimary:
          type: boolean
          default: false
    FinalizeAttachmentRequest:
      type: object
      required:
        - attachmentType
        - fileName
        - mimeType
        - fileSizeBytes
        - r2ObjectKey
      properties:
        attachmentType:
          $ref: '#/components/schemas/AttachmentType'
        propertyId:
          type: string
          format: uuid
          nullable: true
          description: |
            Required for PHOTO/DOCUMENT unless inquiryId is set instead
            (mutually exclusive). Must be absent for RECORDING.
        inquiryId:
          type: string
          format: uuid
          nullable: true
          description: |
            Required for RECORDING. For PHOTO/DOCUMENT, may be used
            instead of propertyId during lightweight inquiry capture
            (Phase 13B) — mutually exclusive with propertyId.
        documentType:
          $ref: '#/components/schemas/DocumentType'
        fileName:
          type: string
        mimeType:
          type: string
        fileSizeBytes:
          type: integer
          minimum: 1
        r2ObjectKey:
          type: string
        isPrimary:
          type: boolean
          default: false
        displayOrder:
          type: integer
          default: 0
    Attachment:
      type: object
      required:
        - id
        - attachmentType
        - fileName
        - mimeType
        - fileSizeBytes
        - r2ObjectKey
      properties:
        id:
          type: string
          format: uuid
        propertyId:
          type: string
          format: uuid
          nullable: true
        inquiryId:
          type: string
          format: uuid
          nullable: true
        attachmentType:
          $ref: '#/components/schemas/AttachmentType'
        documentType:
          $ref: '#/components/schemas/DocumentType'
        fileName:
          type: string
        mimeType:
          type: string
        fileSizeBytes:
          type: integer
          description: Attachment size in bytes.
        r2Bucket:
          type: string
        r2ObjectKey:
          type: string
        fileUrl:
          type: string
          nullable: true
        isPrimary:
          type: boolean
        displayOrder:
          type: integer
        uploadedBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    AttachmentResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Attachment'
    AttachmentListResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
    UploadUrlResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
        data:
          type: object
          required: [uploadUrl, r2ObjectKey, expiresIn]
          properties:
            uploadUrl:
              type: string
              format: uri
            r2ObjectKey:
              type: string
            expiresIn:
              type: integer
              example: 900
    AttachmentDownloadUrlResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
        data:
          type: object
          required: [downloadUrl, expiresIn]
          properties:
            downloadUrl:
              type: string
              format: uri
              description: A presigned R2 GET URL, valid only for expiresIn seconds. Not meant to be stored or reused.
            expiresIn:
              type: integer
              example: 300
...
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                additionalProperties: true
        requestId:
          type: string
...